Inventory UI, Code refinements
+ Got the UI for the inventory working
There are four slots, and the signal "collectItems" is connected to the UI canvas node, which updates the slots based on the sprite path in the inventory JSON
There's also an item counter
+ Code refinements
Got rid of/stopped using the setInfo() function due to how it just completely overwrites the file, so if done wrong, can delete everything. There is of course the setInv() function that uses the setInfo() function to work. The setInv() function is basically just the setInfo() function, but instead of repetitively give all the additional stuff like giving the function the JSON path every time it's called, the setInv() function just takes care of it.
+ Started work on the attack system